Use PackageManager to remove packages
I'd used powershell because the PackageManager.RemovePackageAsync().AsTask() occasionally never completed a valid operation. However using a manual reset event, as in the example code works just fine. To repro the failure with the task code:

1) install WinAppSdk 1.1.3 provisoned
2) install WinAppSdk 1.1.4 for a user

i.e. the provisioned version has a dependency on a non provisioned framework package.

3) attempt to remove the 1.1.4 sdk when running the app non elevated. The task never completes and will always time out.