Make versions.json an embedded resource
This makes the code easier as I don't have to validate its contents, plus if the schema changes the latest version is always available and it simplifies the install. By default the code will now always check for the online version first so I can remove the associated settings too.