Optimize finding sdk packages
Turns out that finding dependents is the bottleneck. Using Paralle.ForEach() on the package list fixed it. I tried paralleling the other loops but it didn't make any difference, neither did using Where() instead of FindAll(). In fact on my middle of the road laptop that was very slightly slower.