Update WinAppSdkCleaner.iss
Add back the finish page, I prefer it.
The uninstall routine no longer works in inno 6.5.4 It seems that the unis000.exe is deleted twice. The second time asynchronously and occasionally after the upgrade was installed. So it deleted the new one from the upgrade. Deleting the private hidden install directory contents first is good enough.